refactor: move ContentSchemaService to Services/ namespace
Follows convention of PedagogyConfigService and SectionProfileService.
Config-loading singletons belong in Services/, not AI/.
Moves test to Tests/Services/ to match.
